refactor: next page 라우터 분기
