+++ src/pages/post/[slug].tsx
@@ -1,105 +1,57 @@
-import dayjs from 'dayjs';
 import { GetStaticPaths, GetStaticProps } from 'next';
 
-import { AnimatedContainer, HoverCard, IconText } from '@/components/common';
-import { CalendarIcon, ListIcon } from '@/components/icons';
-import PostItem from '@/components/post/PostItem';
-import { PageSEO } from '@/components/SEO';
-import {
-  fadeIn,
-  fadeInSlideToLeft,
-  fadeInUp,
-  staggerOne,
-  staggerTwo,
-} from '@/constants/animations';
-import Layout from '@/layouts/Layout';
-import { allSeries, allSeriesName } from '@/libs/dataset';
+import { PostNavigationProps } from '@/components/post/PostNavigation';
+import PostLayout, { PostLayoutProps } from '@/layouts/PostLayout';
+import { allPost, allSeries } from '@/libs/dataset';
+import { parseContents } from '@/libs/mdx';
 import { Series } from '@/types/post';
 
-export const getStaticPaths: GetStaticPaths = () => {
+export const getStaticPaths: GetStaticPaths = async () => {
   return {
-    paths: allSeriesName.map((seriesName) => `/blog/${seriesName}`),
+    paths: allPost.map((post) => post.slug),
     fallback: 'blocking',
   };
 };
 
-export const getStaticProps: GetStaticProps = ({ params }) => {
+export const getStaticProps: GetStaticProps = async ({ params }) => {
   const { slug } = params as { slug: string };
+  const path = `/post/${slug}`;
 
-  const series = allSeries.find((v) => v.seriesName === slug);
+  const post = allPost.find((v) => v.slug === path);
+  const postIndex = allPost.findIndex((v) => v.slug === path);
 
-  if (!series) {
+  if (!post || postIndex < 0) {
     return {
       notFound: true,
     };
   }
 
+  const postNavigation: PostNavigationProps = {
+    prevPost: allPost.at(postIndex + 1) ?? null,
+    nextPost: postIndex === 0 ? null : allPost.at(postIndex - 1) ?? null,
+  };
+
+  let series: Series | null = null;
+
+  if (post.seriesName) {
+    series =
+      allSeries.find((series) =>
+        series._raw.sourceFilePath.startsWith(`series/${post.seriesName}`),
+      ) ?? null;
+  }
+
+  const props: PostLayoutProps = {
+    post,
+    series,
+    postNavigation,
+    tableOfContents: parseContents(post.body.raw),
+  };
+
   return {
-    props: {
-      series,
-    },
+    props,
   };
 };
 
-export default function PostPage({ series }: { series: Series }) {
-  return (
-    <Layout>
-      <PageSEO
-        title={series.title}
-        description={series.description}
-        url={series.slug}
-      />
-
-      <AnimatedContainer variants={staggerTwo} useTransition>
-        <div className="grid gap-8 sm:grid-cols-3 sm:gap-32">
-          <div className="sm:sticky sm:top-8 sm:self-start">
-            <AnimatedContainer
-              variants={fadeInSlideToLeft}
-              className="sm:col-span-1"
-            >
-              <HoverCard>
-                <div className="relative mx-auto h-[336px] w-[240px] select-none rounded-lg bg-neutral-200 px-11 pb-16 pt-12 dark:bg-neutral-800">
-                  <div className="absolute inset-y-0 left-4 w-[1px] bg-neutral-50 dark:bg-neutral-700" />
-                  <div className="text-primary flex h-full break-keep bg-neutral-50 px-3 py-4 text-xl font-semibold dark:bg-neutral-700">
-                    {series.title}
-                  </div>
-                </div>
-              </HoverCard>
-            </AnimatedContainer>
-          </div>
-
-          <div className="sm:col-span-2">
-            <AnimatedContainer
-              variants={fadeIn}
-              className="bg-secondary rounded-lg px-5 py-4"
-            >
-              <p className="text-primary font-medium">{series.description}</p>
-              <div className="text-secondary mt-1 flex gap-2">
-                <IconText
-                  Icon={CalendarIcon}
-                  text={dayjs(series.date).format('YY.MM.DD')}
-                />
-                <IconText Icon={ListIcon} text={`${series.posts.length}편`} />
-              </div>
-            </AnimatedContainer>
-
-            <AnimatedContainer
-              variants={staggerOne}
-              className="mt-16 space-y-4"
-              useTransition
-            >
-              {series.posts.map((post, i) => (
-                <AnimatedContainer key={post.slug} variants={fadeInUp}>
-                  <div className="flex space-x-6">
-                    <div className="pt-4 font-bold">{i + 1}.</div>
-                    <PostItem post={post} />
-                  </div>
-                </AnimatedContainer>
-              ))}
-            </AnimatedContainer>
-          </div>
-        </div>
-      </AnimatedContainer>
-    </Layout>
-  );
+export default function PostPage(props: PostLayoutProps) {
+  return <PostLayout {...props} />;
 }
